Move TUI and banner handling to parser
Start TUI from parser and centralize banner printing there.
Rename CLI flag to "banner" (defaults to true) and pass its
address into buildCmd. Add configCheck in main that fetches the
remote shared config, compares it to local, and prints a colored
warning if they differ.
