+++ cookiefarm/client/cmd/parser.go
@@ -2,14 +2,38 @@ package cmd
 
 import (
 	"context"
+	"logger"
 	"os"
+
+	"client/tui"
 
 	"charm.land/lipgloss/v2"
 	"github.com/charmbracelet/fang"
 )
 
-func ParseArgs(version string, theme fang.ColorScheme, useBanner *bool) {
-	buildCmd(useBanner)
+var (
+	useBanner bool = true
+	useTUI    bool
+)
+
+func startTui() {
+	err := tui.StartTUI(logger.GetBanner("client"))
+	if err != nil {
+		logger.Log.Error().Err(err).Msg("Error starting TUI")
+		if !logger.IsCompletionCommand() {
+			logger.PrintBanner(useBanner, "client")
+		}
+	}
+}
+
+func ParseArgs(version string, theme fang.ColorScheme) {
+	buildCmd(&useBanner)
+
+	if useTUI {
+		startTui()
+	} else if !logger.IsCompletionCommand() {
+		logger.PrintBanner(useBanner, "client")
+	}
 
 	err := fang.Execute(
 		context.TODO(),
